Rebrand, restructure, and upgrade for v2.0.0 release
Major breaking changes: namespace changed to Xrm.Json.Serialization, removed Innofactor branding, updated assembly metadata, and moved projects to root. Upgraded to .NET 4.8, updated Newtonsoft.Json, xUnit, and System.Text.Json. Added changelog, upgrade guide, and restructure script. No changes to serialization logic; all tests and converters updated for new structure.
