+++ Innofactor.Xrm.Json.Serialization.Tests/xunit.runner.json
@@ -1,0 +1,3 @@
+{
+  "methodDisplay": "method"
+}
