--- src/Innofactor.Xrm.Json.Serialization.Tests/packages.config
@@ -1,29 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="10.0.3" targetFramework="net48" />
-  <package id="Microsoft.CrmSdk.CoreAssemblies" version="9.0.2.60" targetFramework="net48" />
-  <package id="Microsoft.IdentityModel" version="7.0.0" targetFramework="net48" />
-  <package id="Microsoft.TestPlatform.ObjectModel" version="18.3.0" targetFramework="net48" />
-  <package id="Newtonsoft.Json" version="13.0.4" targetFramework="net48" />
-  <package id="System.Buffers" version="4.6.1" targetFramework="net48" />
-  <package id="System.Collections.Immutable" version="10.0.3" targetFramework="net48" />
-  <package id="System.IO.Pipelines" version="10.0.3" targetFramework="net48" />
-  <package id="System.Memory" version="4.6.3" targetFramework="net48" />
-  <package id="System.Numerics.Vectors" version="4.6.1" targetFramework="net48" />
-  <package id="System.Reflection.Metadata" version="10.0.3" targetFramework="net48" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.1.2" targetFramework="net48" />
-  <package id="System.ServiceModel.Http" version="10.0.652802" targetFramework="net48" />
-  <package id="System.ServiceModel.Primitives" version="10.0.652802" targetFramework="net48" />
-  <package id="System.Text.Encodings.Web" version="10.0.3" targetFramework="net48" />
-  <package id="System.Text.Json" version="10.0.3" targetFramework="net48" />
-  <package id="System.Threading.Tasks.Extensions" version="4.6.3" targetFramework="net48" />
-  <package id="System.ValueTuple" version="4.6.2" targetFramework="net48" />
-  <package id="xunit" version="2.9.3" targetFramework="net48" />
-  <package id="xunit.abstractions" version="2.0.3" targetFramework="net48" />
-  <package id="xunit.analyzers" version="1.27.0" targetFramework="net48" developmentDependency="true" />
-  <package id="xunit.assert" version="2.9.3" targetFramework="net48" />
-  <package id="xunit.core" version="2.9.3" targetFramework="net48" />
-  <package id="xunit.extensibility.core" version="2.9.3" targetFramework="net48" />
-  <package id="xunit.extensibility.execution" version="2.9.3" targetFramework="net48" />
-  <package id="xunit.runner.visualstudio" version="3.1.5" targetFramework="net48" developmentDependency="true" />
-</packages>
