--- src/Innofactor.Xrm.Json.Serialization.Tests/xunit.runner.json
@@ -1,3 +1,0 @@
-{
-  "methodDisplay": "method"
-}
