--- src/Innofactor.Xrm.Json.Serialization/Properties/AssemblyInfo.cs
@@ -1,36 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Xrm.Json.Serialization")]
-[assembly: AssemblyDescription("Compact JSON serialization for Microsoft Dynamics 365/CRM/Dataverse entities")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Biznamics")]
-[assembly: AssemblyProduct("Xrm.Json.Serialization")]
-[assembly: AssemblyCopyright("Copyright © Biznamics 2025")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("19d18e8d-e967-4cbe-8822-617bd1feeaa4")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("2.0.0.0")]
-[assembly: AssemblyFileVersion("2.0.0.0")]
